Rename module to `storage`
