+++ crates/flipperzero/src/lib.rs
@@ -6,7 +6,7 @@
 extern crate alloc;
 
 pub mod dialogs;
-pub mod filesystem;
 pub mod furi;
 pub mod gui;
 pub mod macros;
+pub mod storage;
