+++ examples/hello-storage/.gitignore
