+++ examples/hello-storage/Cargo.toml
